+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+* 23: Add message field to Reject rules
 * 21: Implement Reject ModRules
 
 ## 0.7.1 - 2020-11-27